kura_nexus/kura/requirements.txt
django
djangorestframework
django-cors-headers
drf-yasg
celery
redis
django-environ
psycopg2-binary